Removing LuaT dependency from lib/TH
